Issue #53: OGP情報追加
