+++ frontend/src/utils/urls.ts
@@ -21,6 +21,10 @@ export const createPostUrl = `${HOST_API_URL}/posts`
 // 投稿取得URL（GET）
 export const getPostByIdUrl = (id: string) => `${HOST_API_URL}/posts/${id}`
 
+// サーバーサイド投稿取得URL（GET）
+export const getServerSidePostByIdUrl = (id: string) =>
+  `${DOCKER_API_URL}/posts/${id}`
+
 // 投稿編集用URL（GET）
 export const editPostUrl = (id: string) => `${HOST_API_URL}/posts/${id}/edit`
